+++ src/index.ts
@@ -13,6 +13,8 @@ import { SaveLogContextMenu } from './save-log/save.contextMenu'
 import { HighlightDecorator } from './highlight/highlight.decorator'
 import { HighlightHotkeyProvider } from './highlight/highlight.hotkeys'
 import { HighlightService } from './highlight/highlight.service'
+import { TabTitleDecorator } from './tab-title/tab-title.decorator'
+import { TabTitleService } from './tab-title/tab-title.service'
 
 @NgModule({
     imports: [
@@ -22,11 +24,13 @@ import { HighlightService } from './highlight/highlight.service'
     ],
     providers: [
         HighlightService,
+        TabTitleService,
         { provide: ConfigProvider, useClass: UartKitConfigProvider, multi: true },
         { provide: HotkeyProvider, useClass: HighlightHotkeyProvider, multi: true },
         { provide: SettingsTabProvider, useClass: UartKitSettingsTabProvider, multi: true },
         { provide: TerminalDecorator, useClass: SaveLogDecorator, multi: true },
         { provide: TerminalDecorator, useClass: HighlightDecorator, multi: true },
+        { provide: TerminalDecorator, useClass: TabTitleDecorator, multi: true },
         { provide: TerminalContextMenuItemProvider, useClass: SaveLogContextMenu, multi: true },
     ],
     entryComponents: [